# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight backend RESTful API designed for tracking and managing personal or organizational expenses. This API provides secure endpoints to log daily expenses, categorize spending, and generate basic financial summaries.",
        "title": "API Backend Expense",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/categories": {
            "get": {
                "description": "get all existing category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "post": {
                "description": "create category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Add new category",
                "parameters": [
                    {
                        "description": "Create new category payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "get one filters by category id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get an category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID (Optional)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "put": {
                "description": "update category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Edit category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Edit category payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCategoryPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Remove category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CreateCategoryPayload": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "dto.Meta": {
            "type": "object",
            "properties": {
                "paginate": {
                    "$ref": "#/definitions/dto.Paginate"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dto.Paginate": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "dto.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apperror.AppError"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/dto.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.UpdateCategoryPayload": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
    }
}